requirements.txt
huggingface_hub>=0.20.0
PyYAML>=6.0
requests>=2.31.0